refactor(server): split mobile routes into modules; align release test with script
- server/routes/mobile.ts (4.1k lines, 76 routes) split into
  server/routes/mobile/ — helpers plus thematic route modules (auth, misc,
  memory, github, git, session, session-lifecycle, workspace, loops) —
  composed with .route("", ...) so every mounted path is byte-identical
  (verified against a pre-split route inventory). Code moved verbatim;
  only imports/exports adjusted, unused imports dropped.
- test/release/automation.test.ts asserted the old combined
  'git push origin HEAD:branch vVersion'; the release script has since
  moved to separate branch push (with rebase-and-retry for
  non-fast-forward races) and tag push — the test now asserts the
  current, safer invariants instead of failing on every run.
