+++ packages/nikcli/src/server/routes/mobile/index.ts
@@ -1,0 +1,24 @@
+import { Hono } from "hono"
+import { lazy } from "@/util/lazy"
+import { AuthRoutes } from "./auth"
+import { MiscRoutes } from "./misc"
+import { MemoryRoutes } from "./memory"
+import { GithubRoutes } from "./github"
+import { SessionRoutes } from "./session"
+import { SessionLifecycleRoutes } from "./session-lifecycle"
+import { WorkspaceRoutes } from "./workspace"
+import { GitRoutes } from "./git"
+import { LoopsRoutes } from "./loops"
+
+export const MobileRoutes = lazy(() =>
+  new Hono()
+    .route("", AuthRoutes())
+    .route("", MiscRoutes())
+    .route("", MemoryRoutes())
+    .route("", GithubRoutes())
+    .route("", SessionRoutes())
+    .route("", SessionLifecycleRoutes())
+    .route("", WorkspaceRoutes())
+    .route("", GitRoutes())
+    .route("", LoopsRoutes()),
+)
